Refactor driver loading to make the public api clearer
